backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg==3.2.10
psycopg-binary==3.2.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
